+++ src/lib/query/block.ts
@@ -1,28 +1,5 @@
 import { graphql } from "lib/gql";
 
-export const getBlockDetailsByHeightQueryDocument = graphql(`
-  query getBlockDetailsByHeight($height: Int!) {
-    blocks_by_pk(height: $height) {
-      hash
-      height
-      timestamp
-      transactions_aggregate {
-        aggregate {
-          sum {
-            gas_used
-            gas_limit
-          }
-        }
-      }
-      validator {
-        moniker
-        operator_address
-        identity
-      }
-    }
-  }
-`);
-
 export const getLatestBlockInfoQueryDocument = graphql(`
   query getLatestBlockInfo {
     blocks(limit: 1, order_by: { height: desc }) {