+++ src/lib/services/accountService.ts
@@ -58,7 +58,7 @@ export const useAccountInfo = (address: Addr): UseQueryResult<AccountInfo> => {
   const endpoint = useBaseApiRoute("accounts");
 
   return useQuery(
-    [CELATONE_QUERY_KEYS.BALANCES, endpoint, address],
+    [CELATONE_QUERY_KEYS.ACCOUNT_DATA, endpoint, address],
     async () => getAccountInfo(endpoint, address),
     { enabled: !!address, retry: 1, refetchOnWindowFocus: false }
   );